changes pageSize to uint64

+++ waku/v2/protocol/waku_store/waku_store.nim
@@ -116,7 +116,7 @@ proc init*(T: type PagingInfo, buffer: seq[byte]): ProtoResult[T] =
   var pagingInfo = PagingInfo()
   let pb = initProtoBuffer(buffer)
 
-  var pageSize: uint32
+  var pageSize: uint64
   discard ? pb.getField(1, pageSize)
   pagingInfo.pageSize = pageSize
 
@@ -247,11 +247,11 @@ proc paginateWithIndex*(list: seq[IndexedWakuMessage], pinfo: PagingInfo): (seq[
     pageSize = pinfo.pageSize
     dir = pinfo.direction
 
-  if pageSize == 0: # pageSize being zero indicates that no pagination is required
+  if pageSize == uint64(0): # pageSize being zero indicates that no pagination is required
     return (list, pinfo)
 
   if list.len == 0: # no pagination is needed for an empty list
-    return (list, PagingInfo(pageSize: 0, cursor:pinfo.cursor, direction: pinfo.direction))
+    return (list, PagingInfo(pageSize: uint64(0), cursor:pinfo.cursor, direction: pinfo.direction))
 
   var msgList = list # makes a copy of the list
   # sorts msgList based on the custom comparison proc indexedWakuMessageComparison
@@ -267,23 +267,23 @@ proc paginateWithIndex*(list: seq[IndexedWakuMessage], pinfo: PagingInfo): (seq[
         cursor = list[list.len - 1].index # perform paging from the end of the list
   var foundIndexOption = msgList.findIndex(cursor) 
   if foundIndexOption.isNone: # the cursor is not valid
-    return (@[], PagingInfo(pageSize: 0, cursor:pinfo.cursor, direction: pinfo.direction))
-  var foundIndex = foundIndexOption.get()
-  var retrievedPageSize, s, e: int
+    return (@[], PagingInfo(pageSize: uint64(0), cursor:pinfo.cursor, direction: pinfo.direction))
+  var foundIndex = uint64(foundIndexOption.get())
+  var retrievedPageSize, s, e: uint64
   var newCursor: Index # to be returned as part of the new paging info
   case dir
     of PagingDirection.FORWARD: # forward pagination
-      let remainingMessages= msgList.len - foundIndex - 1
+      let remainingMessages= uint64(msgList.len) - uint64(foundIndex) - 1
       # the number of queried messages cannot exceed the MaxPageSize and the total remaining messages i.e., msgList.len-foundIndex
-      retrievedPageSize = min(int(pageSize), MaxPageSize).min(remainingMessages)  
+      retrievedPageSize = min(uint64(pageSize), MaxPageSize).min(remainingMessages)  
       if initQuery : foundIndex = foundIndex - 1
       s = foundIndex + 1  # non inclusive
       e = foundIndex + retrievedPageSize 
       newCursor = msgList[e].index # the new cursor points to the end of the page
     of PagingDirection.BACKWARD: # backward pagination
-      let remainingMessages=foundIndex
+      let remainingMessages = foundIndex
       # the number of queried messages cannot exceed the MaxPageSize and the total remaining messages i.e., foundIndex-0
-      retrievedPageSize = min(int(pageSize), MaxPageSize).min(remainingMessages) 
+      retrievedPageSize = min(uint64(pageSize), MaxPageSize).min(remainingMessages) 
       if initQuery : foundIndex = foundIndex + 1
       s = foundIndex - retrievedPageSize 
       e = foundIndex - 1
